% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseModel.R
\name{analyseModel}
\alias{analyseModel}
\title{Analyse a polynomial PDMP}
\usage{
analyseModel(polyModel, model = polyModel, seeds = NULL,
  dir = file.path(getwd(), "simulations"),
  filenameprefix = descr(polyModel), momentorder = c(4, 10),
  plotorder = 1:4, plot = TRUE, modality = TRUE, sim = TRUE,
  momApp = TRUE, lower = NULL, upper = NULL, useCsv = FALSE,
  statistics = c("min", "max", "mean", "median", "sd"))
}
\arguments{
\item{polyModel}{an object of class \code{\link{polyPdmpModel}}}

\item{model}{the corresponding object of class \code{\link{pdmpModel}}. It is
used for simulation because the simulation of an object of class
\code{pdmpModel} is faster than the simulation of an object of class
\code{polyPdmpModel}.}

\item{seeds}{number of seeds to be simulated}

\item{dir}{string giving the directory where all files will be stored.}

\item{filenameprefix}{string. The name of each file saved by \code{analyseModel}
will start with this string.}

\item{momentorder}{numerical vector. Each entry specifies the maximal order
of moments that shall be calculated with \code{\link{momApp}}. The moment
approximation will be performed for each entry seperately. 
Defaults to c(4, 10).}

\item{plotorder}{vector giving all the orders of moments that shall be plotted. 
Defaults to 1:4.}

\item{plot}{boolean variable. Should \code{analyseModel} generate plots?}

\item{modality}{boolean variable. Should \code{analyseModel} test if the 
distribution is unimodal with \code{\link{is.unimodal}}?}

\item{sim}{boolean variable. Should \code{analyseModel} do the simulation
or use already simulated data? In the latter case, it will use files
stored at the same places where \code{analyseModel} would store the simulations.
Furthermore, no value for parameter \code{seeds} is necessary.}

\item{momApp}{boolean variable. Should \code{analyseModel} calculate moments
with \code{\link{momApp}}?}

\item{lower}{integer. Lower bound of the support of the distribution.
This variable is only needed if \code{modality = TRUE}.}

\item{upper}{integer. Upper bound of the support of the distribution.
This variable is only needed if \code{modality = TRUE}.}

\item{useCsv}{boolean variable. Should \code{analyseModel} use 
\code{\link{multSim}} (if \code{FALSE}) or \code{\link{multSimCsv}} 
(if \code{TRUE}) for simulation? Defaults to \code{FALSE}, because
function \code{multSimCsv} is only necessary for huge simulations that 
exceed the working storage.}

\item{statistics}{character vector. Each entry should be the name of a function
that can be applied over simulated data. It will be used by the functions
\code{\link[pdmpsim]{plotStats}} and \code{\link[pdmpsim]{summarise_at}}.}
}
\description{
Perform all methods that are available for a polynomial PDMP:
\itemize{
\item Simulate the PDMP multiple times
\item Plot the simulated data with all available plot methods
\item Calculate statistics such as mean, sd, median, etc
\item Perform moment approximation with all closure methods that are available
and plot the result
\item Test if the distribution is unimodal and plot the result
}
All plots and calculated data will be saved in different files.
It is possible to only perform a part of the methods, i.e. pass already simulated
data and only perform the plot methods or simulate without calculating statistics
or approximated moments.
}
