% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Analyse different polynomial PDMPs at once}
\usage{
analysis(data, model, polyModel, prefix, dir, subdirs, ...)
}
\arguments{
\item{data}{a data.frame, see Details below}

\item{model}{an object of class \code{\link{pdmpModel}}}

\item{polyModel}{the corresponding object of class \code{\link{polyPdmpModel}}}

\item{dir}{string giving the directory where files will be stored.}

\item{...}{additional parameters to function \code{\link{analyseModel}}.}

\item{subDirs}{boolean variable indicating if subdirectories for every
simulation shall be created. Defaults to FALSE.}
}
\description{
Read a csv file which contains informations about the PDMPs
and their simulation parameters. Analyse all PDMPs with
\code{\link{analyseModel}} and save the resulting files in a 
given directory.
}
