% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momApp.R
\name{centralMoment}
\alias{centralMoment}
\title{A help function for \code{\link{momApp}}}
\usage{
centralMoment(lhs, n, p)
}
\arguments{
\item{lhs}{numeric matrix. Each row gives the order of a moment
for which an ode exists.}

\item{n}{numeric: number of all continous variables}

\item{p}{numeric vector giving the order of the moment that shall be centered.
It should have the same length as \code{ncol(lhs)}.}

\item{k}{numeric vector with \code{length(k) = n}.}
}
\description{
Let m be the maximal order for which an entry exists in lhs. 
Let X be a random variable and Y the corresponding centered variable, i.e. \eqn{Y = X - \mu}. 
This function returns the sum 
\deqn{E(X^p) = \sum_{k_1=0}^{p_1}...\sum_{k_n=0}^{p_n} {p \choose k} \mu^{p-k} E(Y^k),}
where \eqn{E(Y^k)} is replaced by an approximated value if it is not found in lhs, or
\deqn{E(Y^k) = \sum_{j_1=0}^{k_1}...\sum_{j_n=0}^{k_n} (-1)^{k-j} {k \choose j} \mu^{k-j} E(X^j).}
}
