% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momApp.R
\name{centralMomentCoef}
\alias{centralMomentCoef}
\title{A help function for \code{\link{centralMoment}}}
\usage{
centralMomentCoef(k, m, p)
}
\arguments{
\item{k}{numeric vector of length n}

\item{m}{Unklar! Gilt m = p? Sind dann die Summen 0?}

\item{p}{numeric vector of length n}
}
\description{
This function computes the sum 
\deqn{\sum_{j_1 = k_1}^m ... \sum_{j_n = k_n}^m (-1)^{j-k} {p \choose j}{j \choose k}.}
The value of \code{centralMomentCoef} is needed in the computations in \code{\link{centralMoment}}.
}
