% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.R
\name{dfactorial}
\alias{dfactorial}
\title{Double factorial}
\usage{
dfactorial(n)
}
\arguments{
\item{n}{a non negative, natural number}
}
\description{
This function returns the double factorial n!! of a natural number n.
It is defined as \eqn{n!! = n*(n-2)*(n-4)*...*1} if n is odd and
\eqn{n!! = n*(n-2)*(n-4)*...*2} if n is even.
}
