% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dtrunc}
\alias{dtrunc}
\title{Probability density function of truncated random variables}
\usage{
dtrunc(x, spec, a = -Inf, b = Inf, ...)
}
\arguments{
\item{x}{ A numeric vector of quantile values }

\item{spec}{ a character value that specifies the underlying probability distribution }

\item{a}{ a numeric value for the lower bound of the random variable }

\item{b}{ a numeric value for the upper bound of the random variable }

\item{...}{other arguments are passed to the corresponding quantile function}
}
\value{
A vector of density function values.
}
\description{
This function computes values for the probability density function of a
truncated random variable. It was originally implemented in package
\pkg{truncdist} and slightly modified to return zeros in case that the
trunction interval [a, b] is not inside the support of the density function.
}
\details{
The \code{R} function is used to access the required probability density and cumulative
  distribution functions of the underlying unrestricted random variable.
}
\examples{
x <- seq(0, 3, 0.1)
dtrunc(x, spec = "norm", a = 1, b = 2)
curve(dtrunc(x, spec = "norm", a = -Inf, b = 1), -10, 2)

\dontrun{
# different results for intervals outside the support of the density function:#' 
truncdist::dtrunc(x, spec = "norm", a = 20, b = 30) # gives error
pdmppoly::dtrunc(x, spec = "norm", a = 20, b = 30) # gives only a warning}
}
\references{
Nadarajah, S. and S. Kotz, 2006. \code{R} Programs for Computing
Truncated Distributions, \emph{Journal of Statistical Software},
16, August 2006, \url{http://www.jstatsoft.org/v16/c02}
}
