% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_existance.R
\name{exists.distribution}
\alias{exists.distribution}
\title{Check for existance of a distribution}
\usage{
exists.distribution(A, B, m)
}
\arguments{
\item{A}{numeric. The lower bound of the support of the distribution.}

\item{B}{numeric. The upper bound of the support of the distriubtion.}

\item{m}{numeric vector giving the non standardized moments \eqn{m_1, m_2,
m_3, ...}{m₁, m₂, m₃}, sorted by their degree.}
}
\value{
boolean variable indicating if a distribution exists (TRUE) or 
not (FALSE). This method does not return a suitable distribution.
}
\description{
This method checks different inequalities to determine if there exists a one
dimensional distribution with given support \eqn{[A, B]} and raw moments
\eqn{m_1, m_2, m_3, ...}{m₁, m₂, m₃, ...}. Only inequalities that include
moments up to order 5 are implemented.
}
