% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypdmp_generator.R
\name{generator}
\alias{generator}
\alias{polyGenerator}
\title{Compute the generator}
\usage{
polyGenerator(obj, ...)
}
\description{
There are several methods to compute the generator for a \code{\link{polyPdmpModel}} object.
}
\note{
Method \code{polyGenerator} only works for one discrete variable.
}
