% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypdmp_methods.R
\name{getIndex}
\alias{getIndex}
\title{get Index of variable 'var' in vector 'vect'}
\usage{
getIndex(var, vect)
}
\arguments{
\item{var}{element that one expects to be part of \code{vect}}

\item{vect}{vector}
}
\description{
This method is used internally. If no index is found with \code{\link{which}}
(this can happen because of numerical issues), the less efficient method
\code{\link{all.equal}} is used to check for "nearly" equality.
}
\examples{
getIndex("c", letters)
getIndex(5.0000001, 2:6) # no result
getIndex(5.00000001, 2:6) # difference is small enough to be found by all.equal
}
