% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_unimodal.R
\name{is.unimodal}
\alias{is.unimodal}
\alias{is.2_b_unimodal}
\alias{is.4_b_unimodal}
\title{Test if moments come from a unimodal distribution with compact support}
\usage{
is.unimodal(A, B, m, eps = 1e-10)

is.2_b_unimodal(A, B, m, eps = 1e-10)

is.4_b_unimodal(A, B, m, eps = 1e-10)
}
\arguments{
\item{A}{numeric. The lower bound of the support of the distribution.}

\item{B}{numeric. The upper bound of the support of the distriubtion.}

\item{m}{numeric vector giving the non standardized moments \eqn{m_1, m_2,
m_3, ...}{m₁, m₂, m₃}, sorted by their degree.}

\item{eps}{numeric value. Some inequalities are of the form \code{... > 0}.
For numerical reasons it is better to test for \code{... > eps} where
\code{eps} is a small number.}
}
\value{
boolean value indicating if the corresponding distribution
is 2-/4-b-unimodal (TRUE) or not (FALSE).
}
\description{
Given a compact support \code{[A, B]} and moments \code{m1, m2, ...} of a 
one dimensional distribution, this method checks if the (unknown) 
distribution can be unimodal. \cr 
There exist several inequalities that test for nonunimodality in the
literature. Depending on the inequality, moments up to order 2 or 4 are
required. A distribution that satisfies all inequalities that contain only
moments up to order 2 is called \emph{2-b-unimodal}. A distribution that
satisfies all inequalities that contain only moments up to order 4 is called
\emph{4-b-unimodal}. The internal methods \code{is.2_b_unimodal} and
\code{is.4_b_unimodal} test these inequalities. Method \code{is.unimodal}
performs these checks, depending on the number of given distributions. A
4-b-unimodal distribution can still be nonunimodal (see examples the below).
But a failure of these tests assures the distribution not to be unimodal.
}
\examples{
### random examples ###

unimodal.example <- function(A=0, B=10){
  dis <- random.distribution(A, B)
  is.unimodal(A, B, mmix(1:4, A, B, distrib = dis))
}
unimodal.example()

### examples that do not work well ###

d1 <- list(list(spec = "unif", min = 3, max = 4),
           list(spec = "lnorm", meanlog = 4))
w1 <- c(10,1)
curve(dmix(a=0, b=10, w1, distrib = d1)(x), 0, 11)
is.unimodal(0, 10, mmix(1:4, 0, 10, w1, distrib = d1))

d2 <- list(list(spec = "unif", min = 6, max = 7),
           list(spec = "norm", mean = 3))
w2 <- c(1,2)
curve(dmix(a=0, b=8, w2, distrib = d2)(x), 0, 8)
is.unimodal(0, 8, mmix(1:4, 0, 8, w2, distrib = d2))

d3 <- list(list(spec = "unif", min = 0, max = 1),
           list(spec = "unif", min = 2, max = 3))
w3 <- c(1,1) # works with c(1,1.2)
B <- 4       # works with B = 3
curve(dmix(a=0, B, w3, distrib = d3)(x), -1, B+1)
is.unimodal(0, B, mmix(1:4, 0, B, w3, distrib = d3))
}
