% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momApp_modalityTest.R
\name{modalityTest}
\alias{modalityTest}
\alias{modalitytest}
\title{Perform modality tests on approximated moments}
\usage{
modalityTest(momApp, lower, upper, vars = names(init(momApp$model)))
}
\arguments{
\item{momApp}{an object of class \code{\link{momApp}}.}

\item{lower}{numeric vector or matrix or data.frame specifying the lower
bounds of the compact distribution that determines the law of the PDMP
\code{momApp$model}. If \code{lower} is a vector, the i-th entry should
give the lower bound of the i-th variable given in \code{vars}, independent
of a time value. If \code{lower} is a matrix or data.frame, it should have
a column named \code{time} containing all time values specified in
\code{times(momApp$model)}. The other column names should be identical to
parameter \code{vars} and contain the lower bounds for the corresponding
variable and time value.}

\item{upper}{numeric vector or matrix or data.frame specifying the upper
bounds of the compact distribution that determines the law of the PDMP
\code{momApp$model}. If \code{upper} is a vector, the i-th entry should
give the upper bound of the i-th variable given in \code{vars}, independent
of a time value. If \code{upper} is a matrix or data.frame, it should have
a column named \code{times} containing all time values specified in
\code{times(momApp$model)}. The other column names should be identical to
parameter \code{vars} and contain the upper bounds for the corresponding
variable and time value.}

\item{vars}{a character vector specifiying the names of the variables
on which a modality test should be performed. The variables should be
part of the PDMP \code{momApp$model}.}
}
\value{
A data.frame containing the test results.
}
\description{
Test if the moments generated by \code{\link{momApp}} 
come from a unimodal distribution with compact support.
The test used is given by function \code{\link[momcalc]{is.unimodal}}
from package \pkg{momcalc}. The lower and upper bounds of
the compact support may depend on a time value \code{t}.
}
\examples{
data(genePolyBF)
support <- getSupport(genePolyBF)
modalityTest(momApp = momApp(genePolyBF), vars = "f",
             lower = support$lower, upper = support$upper)
}
\seealso{
\code{\link{plotModalities}} to plot the result.
}
