% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momApp_outputMethods.R
\name{momApp-methods}
\alias{momApp-methods}
\alias{plot.momApp}
\alias{print.momApp}
\alias{summary.momApp}
\alias{tail.momApp}
\alias{head.momApp}
\title{Methods for objects of class \code{\link{momApp}}}
\usage{
plot.momApp(x, ...)

\method{print}{momApp}(x, ...)

\method{summary}{momApp}(object, ...)

\method{tail}{momApp}(x, ...)

\method{head}{momApp}(x, ...)
}
\arguments{
\item{x}{object of class \code{momApp}}

\item{...}{further arguments to the default method}
}
\description{
Objects of class momApp occur as results of function \code{\link{momApp}}.
Their structure is described in the \code{Details} section of the
documentation of \code{\link{momApp}}. There are several methods to examine
such objects, namely \code{print}, \code{summary}, \code{plot}, \code{tail}
and \code{head}.
}
