% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momApp.R
\docType{methods}
\name{momApp}
\alias{momApp}
\alias{momentApproximation}
\alias{momentapproximation}
\alias{momapp}
\alias{momApp,polyPdmpModel-method}
\title{Moment approximation for polynomial PDMPs}
\usage{
momApp(obj, maxOrder = 4, closure = "zero", centralize = TRUE)

\S4method{momApp}{polyPdmpModel}(obj, maxOrder = 4, closure = "zero",
  centralize = TRUE)
}
\arguments{
\item{obj}{object of class \code{\link{polyPdmpModel}}.}

\item{maxOrder}{integer defining the highest order of moments that are considered.
Higher degrees are droped and replaced by other values. The replacement
method is specified in parameter \code{closure}.}

\item{closure}{string defining the method that does the moment closure, i. e.
that changes the system of ODEs into a closed form that is solvable.
Possible values are \code{setZero} (the default) and reduceDegree.}
}
\value{
an object of class \code{momApp} (see details).
}
\description{
Moment approximation for polynomial PDMPs
}
\details{
The returned s3 class \code{momApp} contains 6 different elements:
\itemize{
\item \code{model}: the polyPdmpModel \code{obj}
\item \code{discRes}: a matrix giving the calculated moments of the different
indicator variables that replace the discrete variable (see \code{\link{blowupSpray}}
for an explanation of the indicator variables)
\item \code{contRes}: a matrix giving the calculated moments of the continous variables
\item \code{maxOrder}: integer defining the highest order of moments to be calculated
\item \code{closure}: string giving the closure method. See ... for more details.
\item \code{contInd}: a data.frame with all moment indexes that are calculated
\item \code{moments}: a data.frame with the resulting moments, of the same structure
 as the result of function \code{\link[pdmpsim]{moments}} in package \pkg{pdmpsim}.
}
}
\note{
This method works only for PDMPs with one discrete variable.
}
\examples{
data(genePolyK2)
a <- momApp(genePolyK2, maxOrder = 4)
plot(a)
print(a)
summary(a)
}
\seealso{
\code{\link{momentClosure}} for the internal method that performs
  the moment closure, \code{\link{momApp-methods}} for further analysing
  the result.
}
