% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momentApproximation.R
\docType{methods}
\name{momentApprox}
\alias{momentApprox}
\alias{momentApproximation}
\alias{momentapproximation}
\alias{momentapprox}
\alias{momentApprox,polyPdmpModel-method}
\title{Moment approximation for polynomial PDMPs}
\usage{
momentApprox(obj, l = 4, closure = "setZero", plot = TRUE, ...)

\S4method{momentApprox}{polyPdmpModel}(obj, l = 4, closure = "setZero",
  plot = TRUE)
}
\arguments{
\item{obj}{object of class \code{\link{polyPdmpModel}}.}

\item{l}{integer defining the highest degree of moments that are considered.
Higher degrees are droped and replaced by other values. The replacement
method is specified in parameter \code{closure}.}

\item{closure}{string defining the method that does the moment closure, i. e.
that changes the system of ODEs into a closed form that is solvable.
Possible values are \code{setZero} (the default) and reduceDegree.}

\item{plot}{boolean value indicating if the results shall be plotted.
Defaults to TRUE.}
}
\value{
an object of class \code{\link{momApp}}.
}
\description{
Moment approximation for polynomial PDMPs
}
\seealso{
\code{link{momentClosure}} for the internal method that performs 
  the moment closure.
}
