% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momApp.R
\name{momentClosure}
\alias{momentClosure}
\title{#' @importFrom spray as.spray
#' @export
momentClosure <- function(name, ode, l, n){
  value <- value(ode)
  index <- index(ode)
  problematicRows <- which(rowSums(index) > l)
  for(i in problematicRows){
    switch(name,
           setZero = {value[i] <- 0},
           reduceDegree = {
             index[i, 1:n] <- sapply(index[i, 1:n], function(j) max(0, j-1))
             return(momentClosure(name, as.spray(index, value, addrepeats = TRUE), l, n))
             },
           stop("moment closure method \"", name, "\" is not implementet.")
    )
  }
  return(as.spray(index, value, addrepeats = TRUE))
}}
\usage{
momentClosure(odeList, lhs, centralize = T, n)
}
\arguments{
\item{odeList}{list of spray objects. Every entry corresponds to a differential equation.}

\item{lhs}{a matrix with nrow(lhs) =  length(odeList). Row i of lhs represents the spray object
f = product(t[i, ]). It stands for the left hand side of the ode E(f(t))' which is stored in 
odeList[[i]].}

\item{centralize}{boolean variable indicating if centralized moments should be replaced instead
of raw moments}

\item{n}{number of continous variables}
}
\description{
#' @importFrom spray as.spray
#' @export
momentClosure <- function(name, ode, l, n){
  value <- value(ode)
  index <- index(ode)
  problematicRows <- which(rowSums(index) > l)
  for(i in problematicRows){
    switch(name,
           setZero = {value[i] <- 0},
           reduceDegree = {
             index[i, 1:n] <- sapply(index[i, 1:n], function(j) max(0, j-1))
             return(momentClosure(name, as.spray(index, value, addrepeats = TRUE), l, n))
             },
           stop("moment closure method \"", name, "\" is not implementet.")
    )
  }
  return(as.spray(index, value, addrepeats = TRUE))
}
}
