% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{mtrunc}
\alias{mtrunc}
\title{Moments of truncated random variables}
\usage{
mtrunc(order, spec, a = -Inf, b = Inf, ...)
}
\arguments{
\item{order}{numeric vector giving the order of the moments}

\item{spec}{ a character value that specifies the underlying probability distribution }

\item{a}{ a numeric value for the lower bound of the random variable }

\item{b}{ a numeric value for the upper bound of the random variable }

\item{...}{other arguments are passed to the corresponding moment function}
}
\description{
This function computes the raw moments of a truncated random variable.
}
\examples{
mtrunc(1:6, spec = "norm")
mtrunc(1:6, spec = "norm", b = 0)
}
\seealso{
\code{\link{dtrunc}} for the probability distribution of a truncated variable.
}
