% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combinatorics.R
\name{multinomial}
\alias{multinomial}
\title{Multinomial coefficient}
\usage{
multinomial(m, k)
}
\arguments{
\item{m}{a non negative, natural number}

\item{k}{a vector of non negative, natural numbers}
}
\description{
This function returns the multinomial coefficient of a natural number m
and a vector of natrual numbers k.
It is defined as \eqn{m!\(k[1]!*...*k[n]!)}, 
where \code{n = length(k)} and \eqn{m!} stands for \code{factorial(m)}.
}
