% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spray_extension.R
\name{my.subs}
\alias{my.subs}
\title{#' @importFrom spray spraymaker
my.spray <- function (M, x, addrepeats = FALSE){
  if (is.null(M)) { #new
    return(spraymaker(list(NULL, NULL)))
  }
  if (is.vector(M)) {
    M <- rbind(M)
  }
  if (inherits(M, "partition")) {
    M <- t(M)
  }
  M <- as.matrix(M)
  if (missing(x)) {
    x <- rep(1, nrow(M))
  }
  if (length(x) == 1) {
    x <- rep(x, nrow(M))
  }
  return(spraymaker(list(M, x), addrepeats = addrepeats))
}}
\usage{
my.subs(S, dims, x, keepArity = FALSE)
}
\description{
# assign to "spray"
unlockBinding("spray", as.environment("package:spray"))
assignInNamespace("spray", my.spray, ns="spray", envir=as.environment("package:spray"))
assign("spray", my.spray, as.environment("package:spray"))
lockBinding("spray", as.environment("package:spray"))
}
