% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momApp_modalityTest.R
\name{plotModalities}
\alias{plotModalities}
\alias{plotmodalities}
\alias{plotModality}
\alias{plotmodality}
\title{Plot the result of \code{\link{modalityTest}}}
\usage{
plotModalities(momApp, modalities = NULL,
  vars = names(init(momApp$model)), ...)
}
\arguments{
\item{momApp}{an object of class \code{\link{momApp}}.}

\item{modalities}{The result of \code{modalityTest(momApp, ...)}. If
\code{modalities = NULL}, function \code{\link{modalityTest}} will be
called inside \code{plotModalities}.}

\item{vars}{character vector with variable names. Only the test results for
these variables will appear in the plot. Vector \code{vars} should be equal
or a subset to \code{names(init(momApp$model))}.}

\item{...}{additional arguments to \code{\link{modalityTest}}. They are
only used if \code{modalities = NULL}.}
}
\value{
an object of class \code{ggplot}
}
\description{
Test if the moments generated by \code{\link{momApp}} 
come from a unimodal distribution with compact support
and plot the result. The test is done by function 
\code{\link{modalityTest}}. Already existing test results
can be passed via the argument \code{modalities}.
}
