% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypdmp_accessors.R
\docType{methods}
\name{polypdmp-accessors}
\alias{polypdmp-accessors}
\alias{dynpolys}
\alias{dynpolys<-}
\alias{ratepolys}
\alias{ratepolys<-}
\alias{ratesprays}
\alias{ratesprays<-}
\alias{dynsprays}
\alias{dynsprays<-}
\alias{dynpolys,polyPdmpModel-method}
\alias{ratepolys,polyPdmpModel-method}
\alias{ratesprays,polyPdmpModel-method}
\alias{dynsprays,polyPdmpModel-method}
\alias{dynpolys<-,polyPdmpModel-method}
\alias{ratepolys<-,polyPdmpModel-method}
\alias{parms<-,polyPdmpModel-method}
\alias{dynfunc<-,polyPdmpModel-method}
\alias{ratefunc<-,polyPdmpModel-method}
\alias{ratesprays<-,polyPdmpModel-method}
\alias{dynsprays<-,polyPdmpModel-method}
\title{Accessor functions for Class polyPdmpModel}
\usage{
dynpolys(obj)

dynpolys(obj) <- value

ratepolys(obj)

ratepolys(obj) <- value

ratesprays(obj)

ratesprays(obj) <- value

dynsprays(obj)

dynsprays(obj) <- value

\S4method{dynpolys}{polyPdmpModel}(obj)

\S4method{ratepolys}{polyPdmpModel}(obj)

\S4method{ratesprays}{polyPdmpModel}(obj)

\S4method{dynsprays}{polyPdmpModel}(obj)

\S4method{dynpolys}{polyPdmpModel}(obj) <- value

\S4method{ratepolys}{polyPdmpModel}(obj) <- value

\S4method{parms}{polyPdmpModel}(obj) <- value

\S4method{dynfunc}{polyPdmpModel}(obj) <- value

\S4method{ratefunc}{polyPdmpModel}(obj) <- value

\S4method{ratesprays}{polyPdmpModel}(obj) <- value

\S4method{dynsprays}{polyPdmpModel}(obj) <- value
}
\arguments{
\item{obj}{an object of class \code{\link{polyPdmpModel}}}

\item{value}{the value that shall be set}
}
\description{
Accessor functions for Class polyPdmpModel
}
\note{
only works for one discrete variable

only works for one discrete variable
}
