% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{random.distribution}
\alias{random.distribution}
\title{Create and plot a random distribution}
\usage{
random.distribution(lower = 0, upper = 10, plot = TRUE)
}
\arguments{
\item{lower}{lower bound of the support of the distribution. Defaults to 0.}

\item{upper}{upper bound of the support of the distribution. Defaults to 10.}

\item{plot}{boolean variable. Should the distribution be plotted?}
}
\value{
A list with the distributions that are components of the mixture
  distribution. It has the same structure as parameter \code{distrib} in
  function \code{\link{dmix}}.
}
\description{
This function computes a random truncated distribution which is a mixture of
one up to 4 different distributions that can be uniform, exponential,
normal or lognormal distributed. It can be used to test the evectiveness
of function \code{\link{is.unimodal}}.
}
\examples{
par(mfrow = c(1, 2))
d <- random.distribution(plot = TRUE)
curve(dmix(0, 10, distrib = d)(x), -1, 10, type = "l", col = "red")
}
