% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypdmp_methods.R
\name{ratespraysToMatrix}
\alias{ratespraysToMatrix}
\title{Convert ratesprays into a matrix}
\usage{
ratespraysToMatrix(obj)
}
\arguments{
\item{obj}{object of class \code{\link{polyPdmpModel}}.}
}
\value{
a two dimensional list of \code{spray} objects.
}
\description{
This method computes a 'ratematrix' out of slot \code{ratesprays}.
The i,j - entry will then be accessible via \code{ratematrix[[i]][[j]]}
and contain the rate from discrete state \code{i} to discrete state \code{j}.
Note that the returned object is not of class \code{matrix}, but is a
nested list of \code{spray} objects.
}
\note{
this method only works for one discrete variable
}
