% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypdmp_methods.R
\name{redefineDynpolys}
\alias{redefineDynpolys}
\title{Redefine Slot \code{dynpolys}}
\usage{
redefineDynpolys(x, obj, where = NULL, overall = 0)
}
\arguments{
\item{x}{an object of type language as described in section Dynpolys in
\code{\link{polyPdmpModel}}.}

\item{obj}{an object of class polyPdmpModel.}

\item{where}{this parameter is necessary because \code{redefineDynpolys} 
is a recursive function. Changing it will have no effect on the result.}

\item{overall}{this parameter is necessary because \code{redefineDynpolys} 
is a recursive function. Please do not change its value!}
}
\value{
a modified object of type language as described in section 
Dynpolys in \code{\link{polyPdmpModel}}.
}
\description{
This function is used internally for function \code{dynpolys<-} that defines
slot \code{dynpolys} and during initialization of a
\code{\link{polyPdmpModel}} object. It replaces parameters stored in slot
\code{parms} by its values and coerces all numbers to \code{spray} objects.
It coerces all list elements that are given in variant 2 to spray objects
defined as in variant 1 (see section Dynpolys in \code{\link{polyPdmpModel}}
for a description of the two variants).
}
\note{
This function only works for one discrete variable
}
