% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypdmp_methods.R
\name{redefineRatepolys}
\alias{redefineRatepolys}
\title{Redefine Slot \code{ratepolys}}
\usage{
redefineRatepolys(x, obj, where = NULL)
}
\arguments{
\item{x}{an object of type language as described in section Ratepolys in
\code{\link{polyPdmpModel}}.}

\item{obj}{an object of class polyPdmpModel.}

\item{where}{this parameter is necessary because \code{redefineRatepolys} 
is a recursive function. Changing it will have no effect on the result.}
}
\value{
an modified object of type language as described in section 
Ratepolys in \code{\link{polyPdmpModel}}.
}
\description{
This function is used internally for function \code{ratepolys<-} that 
defines slot \code{ratepolys} and during initialization of a
\code{\link{polyPdmpModel}} object. It replaces parameters stored in slot
\code{parms} by its values and coerces all numbers to \code{spray} objects.
}
