% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simplePoly}
\alias{simplePoly}
\title{Model \code{simplePoly}}
\format{An object of class \code{polyPdmpModel} of length 1.}
\usage{
simplePoly
}
\description{
This model is equivalent to model \code{\link{simplePdmp}}, which is given as
example in package \pkg{pdmpsim}. Model \code{simplePoly} is defined as a
\code{\link{polyPdmpModel}} object. Apart from that, it models exactly the
same mechanism as \code{simplePdmp}. It is included to the package for
demonstration purposes and is used in some unit tests and function examples.
}
\examples{
## the code used to generate this model:
simplePoly <- polyPdmpModel(
   descr = "polyModel with two jumptypes",
   init = c(f = 0, d = 0),
   times = c(from = 0, to = 10, by = 0.01),
   discStates = list(d = -1:1),
   dynpolys = quote(list(
     list(overall = lone(2,2)) # variant 2
   )),
   ratepolys = quote(list(
     list(0, 1, 2), # first jumptype
     list(2, 1, 0)  # second jumptype
   )),
   jumpfunc = function(t, x, parms, jtype){
     c(0, switch(jtype, x["d"]-1, x["d"]+1))
   }
 )
 
## load it and compare to simplePdmp:
data("simplePoly")
data("simplePdmp")
identical(sim(simplePoly, outSlot = FALSE, seed = 5),
          sim(simplePdmp, outSlot = FALSE, seed = 5))
}
\keyword{datasets}
