% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supports.R
\name{support}
\alias{support}
\alias{getSupport}
\alias{supports}
\alias{plotSupport}
\alias{getsupport}
\alias{plotsupport}
\title{Return the support for gene regulation models}
\usage{
getSupport(model)
}
\arguments{
\item{model}{Object of class \code{pdmpModel} or \code{polyPdmp}. This has to
be one of the previous mentioned models.}
}
\value{
A list with two entries named 'lower' and 'upper'. Both are data.frames
containing the values of the lower / upper bound for each variable of the PDMP
and each time value given in \code{times(model)}.
}
\description{
This package contains multiple PDMPs that simulate gene regulation
mechanisms. They can be loaded with \code{data(model)} where \code{model}
is one of the following: 
\itemize{
\item \code{\link{genePdmpK}}: gene regulation with constant activation
\item \code{\link{genePdmpK2}}: gene regulation with constant activation, 
transcription and translation are modeled seperately
\item \code{\link{genePdmpF}}: gene regulation with positive feedback
\item \code{\link{genePdmpBF}}: gene regulation with positive feedback
and basal transcription rate
\item \code{\link{genePdmpKF}}: gene regulation with constant activation
and positive feedback
\item \code{\link{toggleSwitch}}: toggle switch for two genes
} 
The distribution of each of these models has a compact support (see References
for further details). Method \code{getSupport} returns a list with two data.frames 
containing the lower and upper bounds of the support. They are time-depend 
and are given for every variable of the model seperately.
}
\examples{
data(genePolyT)
getSupport(genePolyT)

}
