% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyseModel.R
\name{analyseModel}
\alias{analyseModel}
\title{Analyse a polynomial PDMP}
\usage{
analyseModel(polyModel, model = polyModel, seeds = NULL,
  dir = file.path(getwd(), "simulations"),
  filenameprefix = descr(polyModel), momentorder = c(4, 10),
  plotorder = 1:4, plot = TRUE, modality = TRUE, sim = TRUE,
  statistics = TRUE, momApp = TRUE, lower = NULL, upper = NULL,
  funs = c("min", "max", "mean", "median", "sd"), title = descr(model))
}
\arguments{
\item{polyModel}{an object of class \code{\link{polyPdmpModel}}}

\item{model}{the corresponding object of class \code{\link{pdmpModel}}. It is
used for simulation because the simulation of an object of class
\code{pdmpModel} is faster than the simulation of an object of class
\code{polyPdmpModel}.}

\item{seeds}{number of seeds to be simulated}

\item{dir}{string giving the directory where all files will be stored.}

\item{filenameprefix}{string. The name of each file saved by \code{analyseModel}
will start with this string.}

\item{momentorder}{numerical vector. Each entry specifies the maximal order
of moments that shall be calculated with \code{\link{momApp}}. The moment
approximation will be performed for each entry seperately. 
Defaults to c(4, 10).}

\item{plotorder}{vector giving all the orders of moments that shall be plotted. 
Defaults to 1:4.}

\item{plot}{boolean variable. Should \code{analyseModel} generate plots?}

\item{modality}{boolean variable. Should \code{analyseModel} test if the 
distribution is unimodal with \code{\link{is.unimodal}}?}

\item{sim}{boolean variable. Should \code{analyseModel} do the simulation
or use already simulated data? In the latter case, it will use files
stored at the same places where \code{analyseModel} would store the simulations.
Furthermore, no value for parameter \code{seeds} is necessary.}

\item{statistics}{boolean variable. Should \code{analyseModel} calculate statistics
with \code{\link{summarise_at}}? The functions that shall be applied on the data
can be specified with the argument \code{funs}.}

\item{momApp}{boolean variable. Should \code{analyseModel} calculate moments
with \code{\link{momApp}}?}

\item{lower}{numeric vector or matrix or data.frame specifying the lower
bounds of the compact distribution that determines the law of the PDMP. #'
It is an argument to function \code{\link{modalityTest}} and only needed if
\code{modality = TRUE}. If \code{upper} . If \code{lower} is a vector, the
i-th entry should give the lower bound of the i-th continous variable given
in \code{init(model)}, independent of a time value. If \code{lower} is a
matrix or data.frame, it should have a column named \code{time} containing
all time values specified in \code{times(model)}. The other column names
should be identical to the continous variables of the PDMP and contain the
lower bounds for the corresponding variable and time value.}

\item{upper}{numeric vector or matrix or data.frame specifying the upper
bounds of the compact distribution that determines the law of the PDMP. It
is an argument to function \code{\link{modalityTest}} and only needed if
\code{modality = TRUE}. If \code{upper} is a vector, the i-th entry should
give the upper bound of the i-th continous variable given in
\code{init(model)}, independent of a time value. If \code{upper} is a
matrix or data.frame, it should have a column named \code{time} containing
all time values specified in \code{times(model)}. The other column names
should be identical to the continous variables of the PDMP and contain the
upper bounds for the corresponding variable and time value.}

\item{funs}{character vector. Each entry should be the name of a function
that can be applied over simulated data. It will be used by the functions
\code{\link[pdmpsim]{plotStats}} and \code{\link[pdmpsim]{summarise_at}}.}
}
\description{
Perform all methods that are available for a polynomial PDMP:
\itemize{
\item Simulate the PDMP multiple times
\item Plot the simulated data with all available plot methods
\item Calculate statistics such as mean, sd, median, etc
\item Perform moment approximation with all closure methods that are available
and plot the result
\item Test if the distribution is unimodal and plot the result
}
All plots and calculated data will be saved in different files.
It is possible to only perform a part of the methods, i.e. pass already simulated
data and only perform the plot methods or simulate without calculating statistics
or approximated moments.
}
