% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdmp_analysis.R
\name{analysis}
\alias{analysis}
\title{Analysis of models used in thesis ...}
\usage{
analysis(data, model, seeds = 1:50, useCsv = F, dir = getwd())
}
\arguments{
\item{data}{a data.frame, see Details below}

\item{model}{an object of class \code{\link{pdmpModel}}}

\item{seeds}{number of seeds to be simulated}

\item{useCsv}{boolean variable, indicating if \code{\link{multSim}} (FALSE)
or \code{\link{multSimCsv}} should be used for simulation. Defaults to FALSE.}

\item{dir}{string giving the directory where files will be stored.
Defaults to the working directory.}
}
\description{
Analysis of models used in thesis ...
}
