% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis.R
\name{analysis}
\alias{analysis}
\title{Analysis of models used in PROM}
\usage{
analysis(data, model, polyModel, seeds = 1:50, useCsv = FALSE,
  dir = file.path(getwd(), "simulations"), subDirs = FALSE,
  momentorders = 1:10, plotorder = 1:4, plot = TRUE, sim = TRUE,
  lower = NULL, upper = NULL)
}
\arguments{
\item{data}{a data.frame, see Details below}

\item{model}{an object of class \code{\link{pdmpModel}}}

\item{polyModel}{the corresponding object of class \code{\link{polyPdmpModel}}}

\item{seeds}{number of seeds to be simulated}

\item{useCsv}{boolean variable indicating if \code{\link{multSim}} (if FALSE)
or \code{\link{multSimCsv}} (if TRUE) should be used for simulation.}

\item{dir}{string giving the directory where files will be stored.}

\item{subDirs}{boolean variable indicating if subdirectories for every
simulation shall be created. Defaults to FALSE.}

\item{momentorders}{vector giving all the orders of moments that shall 
be calculated. Defaults to 1:10.}

\item{plotorder}{vector giving all the orders of moments that shall be plotted.
Defaults to 1:4.}

\item{plot}{boolean variable. Should \code{analysis} generate plots?}

\item{sim}{boolean variable. Should \code{analysis} do the simulation
or use already simulated data? In the latter case, it will use files
stored at the same places where \code{analysis} would store the simulations.}

\item{lower}{integer. Lower bound of the compact support of the distribution.}

\item{upper}{integer. Upper bound of the compact support of the distribution.}
}
\description{
Analysis of models used in PROM
}
