% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dmix}
\alias{dmix}
\alias{mmix}
\title{Mixtures of truncated distributions}
\usage{
dmix(a = -Inf, b = Inf, distrib, weights)
}
\arguments{
\item{a}{numeric giving the lower bound of the support. Defaults to -Inf.}

\item{b}{numeric giving the upper bound of the support. Defaults to Inf.}

\item{weights}{numeric vector with the same length as \code{distrib}.
Provides weights for every distribution given in \code{distrib}.}
}
\description{
Mixtures of truncated distributions
}
\examples{
distributions <- list(list(spec="exp", rate = 2),
                     list(spec="norm", mean=0, sd = 0.5),
                     list(spec="unif", min=2, max=3))
curve(dmix(3, a=-1, weights = c(.1,.3,.1), distrib=distributions)(x),-2,5)
}
