% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions.R
\name{dmix}
\alias{dmix}
\alias{mmix}
\title{Mixtures of truncated distributions}
\usage{
dmix(lower = -Inf, upper = Inf, distrib, weights)

mmix(order, lower = -Inf, upper = Inf, distrib, weights)
}
\arguments{
\item{lower}{numeric giving the lower bound of the support. Defaults to -Inf.}

\item{upper}{numeric giving the upper bound of the support. Defaults to Inf.}

\item{distrib}{a list. Every element is itself a list representing a
distribution. This list should have one named element \code{spec} which
describes the distribution, i. e. "exp" for the exponential distribution or
"norm" for the normal distribution. The other elements are optional
additional parameters for the specfied distribution.}

\item{weights}{numeric vector with the same length as \code{distrib}.
Provides weights for every distribution given in \code{distrib}.}

\item{order}{order of the moment}
}
\description{
Compute the distribution (\code{dmix}) and raw moments (\code{mmix}) of a
mixture of different, eventually truncated distributions.
}
\examples{
distributions <- list(list(spec="exp", rate = 2),
                     list(spec="norm", mean = 0, sd = 0.5),
                     list(spec="unif", min = 2, max = 3))
d <- dmix(-1, 3, weights = c(.1,.3,.1), distrib = distributions)
curve(d(x), -2, 5)
}
\seealso{
\code{\link{random.distribution}} to plot a randomly created mixture
of truncated distributions.
}
