% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_existance.R
\name{exists.distribution}
\alias{exists.distribution}
\title{Check for existance of a distribution}
\usage{
exists.distribution(lower, upper, moments)
}
\arguments{
\item{lower}{numeric. The lower bound A of the support \eqn{[A, B]} of the distribution.}

\item{upper}{numeric. The upper bound B of the support \eqn{[A, B]} of the distriubtion.}

\item{moments}{numeric vector giving the non standardized moments \eqn{m_1, m_2,
m_3, ...}{m₁, m₂, m₃, ...}, sorted by their degree. 
This vector should have at least two entries.}
}
\value{
boolean variable indicating if a distribution exists (TRUE) or 
not (FALSE). This method does not return a suitable distribution.
}
\description{
This method checks different inequalities to determine if there exists a one
dimensional distribution with given support \eqn{[A, B]} and raw moments
\eqn{m_1, m_2, m_3, ...}{m₁, m₂, m₃, ...}. Only inequalities that include
moments up to order 5 are implemented.
}
\examples{
exists.distribution(0, 1, actuar::mbeta(1:4, 0.5, 0.5)) # TRUE
exists.distribution(-1, 2, actuar::munif(1:4, min = -1, max = 2)) # TRUE
exists.distribution(-4, -5, 1:5) # FALSE
}
