% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distributions_unimodal.R
\name{is.unimodal}
\alias{is.unimodal}
\alias{is.2_b_unimodal}
\alias{is.4_b_unimodal}
\title{Test if moments come from a unimodal distribution with compact support}
\usage{
is.unimodal(lower, upper, moments, eps = 1e-10)

is.2_b_unimodal(lower, upper, moments, eps = 1e-10)

is.4_b_unimodal(lower, upper, moments, eps = 1e-10)
}
\arguments{
\item{lower}{numeric. The lower bound A of the support \eqn{[A, B]} of the distribution.}

\item{upper}{numeric. The upper bound B of the support \eqn{[A, B]} of the distriubtion.}

\item{moments}{numeric vector giving the non standardized moments \eqn{m_1, m_2,
m_3, ...}{m₁, m₂, m₃, ...}, sorted by their degree. 
This vector should have at least two entries.}

\item{eps}{numeric value. Some inequalities are of the form \code{... > 0}.
For numerical reasons it is better to test for \code{... > eps} where
\code{eps} is a small number.}
}
\value{
boolean value indicating if the corresponding distribution
is 2-/4-b-unimodal (TRUE) or not (FALSE).
}
\description{
Given a compact support \code{[A, B]} and moments \code{m1, m2, ...} of a 
one dimensional distribution, this method checks if the (unknown) 
distribution can be unimodal. \cr 
There exist several inequalities that test for nonunimodality in the
literature. Depending on the inequality, moments up to order 2 or 4 are
required. A distribution that satisfies all inequalities that contain only
moments up to order 2 is called \emph{2-b-unimodal}. A distribution that
satisfies all inequalities that contain only moments up to order 4 is called
\emph{4-b-unimodal}. The internal methods \code{is.2_b_unimodal} and
\code{is.4_b_unimodal} test these inequalities. Method \code{is.unimodal}
performs these checks, depending on the number of given distributions. A
4-b-unimodal distribution can still be nonunimodal (see the examples below).
But a failure of these tests assures the distribution not to be unimodal.
}
\examples{
### random example ###

  unimodal.example <- function(A = 0, B = 10){
    dis <- random.distribution(A, B)
    is.unimodal(A, B, mmix(1:4, A, B, distrib = dis))
  }
  unimodal.example()


### mixture of two normal distributions ###

  d <- list(list("norm", mean = 3),
            list("norm", mean = 13))
  
  # recognized as not unimodal:
  is.unimodal(0, 20, mmix(a = 0, b = 20, distrib = d, weights = c(0.3, 0.7)))
  # bimodal but not recognized as such:
  is.unimodal(0, 20, mmix(a = 0, b = 20, distrib = d))
  
  
  # bimodal but not recognized as such:
  d[[2]]["mean"] <- 8
  is.unimodal(0, 20, mmix(a = 0, b = 20, distrib = d, weights = c(0.3, 0.7)))
  

### examples that do not work well ###

  # very close to unimodal:
  d1 <- list(list(spec = "unif", min = 3, max = 4),
             list(spec = "lnorm", meanlog = 4))
  w1 <- c(10,1)
  curve(dmix(a = 0, b = 10, weights = w1, distrib = d1)(x), 0, 11)
  is.unimodal(0, 10, mmix(1:4, 0, 10, weights = w1, distrib = d1)) # FALSE

  # bimodal but not recognized as such:
  d2 <- list(list(spec = "unif", min = 6, max = 7),
             list(spec = "norm", mean = 3))
  w2 <- c(1,2)
  curve(dmix(a = 0, b = 8, weights = w2, distrib = d2)(x), 0, 8)
  is.unimodal(0, 8, mmix(1:4, 0, 8, weights = w2, distrib = d2)) # TRUE

}
