% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/momApp.R
\docType{methods}
\name{momApp}
\alias{momApp}
\alias{momentApproximation}
\alias{momentapproximation}
\alias{momapp}
\alias{momApp,polyPdmpModel-method}
\title{Moment approximation for polynomial PDMPs}
\usage{
momApp(obj, l = 4, closure = "setZero")

\S4method{momApp}{polyPdmpModel}(obj, l = 4, closure = "setZero")
}
\arguments{
\item{obj}{object of class \code{\link{polyPdmpModel}}.}

\item{l}{integer defining the highest degree of moments that are considered.
Higher degrees are droped and replaced by other values. The replacement
method is specified in parameter \code{closure}.}

\item{closure}{string defining the method that does the moment closure, i. e.
that changes the system of ODEs into a closed form that is solvable.
Possible values are \code{setZero} (the default) and reduceDegree.}
}
\value{
an object of class \code{\link{momApp}}.
}
\description{
Moment approximation for polynomial PDMPs
}
\details{
The returned s3 class \code{momApp} contains 6 different elements:
\itemize{
\item \code{model}: the polyPdmpModel \code{obj}
\item \code{discRes}: a matrix giving the calculated moments of the different
indicator variables that replace the discrete variable (see \code{\link{blowupSpray}}
for an explanation of the indicator variables)
\item \code{contRes}: a matrix giving the calculated moments of the continous variables
\item \code{degree}: integer defining the highest degree of moments to be calculated
\item \code{closure}: string giving the closure method. See ... for more details.
\item \code{contInd}: a data.frame with all moment indexes that are calculated
\item \code{moments}: a data.frame with the resulting moments, of the same structure
 as \code{\link[pdmpsim]moments}.
}
}
\note{
This method works only for PDMPs with one discrete variable.
}
\seealso{
\code{\link{momentClosure}} for the internal method that performs
  the moment closure.
}
