% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polypdmp_accessors.R
\docType{methods}
\name{polypdmp-accessors}
\alias{polypdmp-accessors}
\alias{dynpolys}
\alias{dynpolys<-}
\alias{ratepolys}
\alias{ratepolys<-}
\alias{ratesprays}
\alias{ratesprays<-}
\alias{dynsprays}
\alias{dynsprays<-}
\alias{dynpolys,polyPdmpModel-method}
\alias{ratepolys,polyPdmpModel-method}
\alias{ratesprays,polyPdmpModel-method}
\alias{dynsprays,polyPdmpModel-method}
\alias{dynpolys<-,polyPdmpModel-method}
\alias{ratepolys<-,polyPdmpModel-method}
\alias{parms<-,polyPdmpModel-method}
\alias{dynfunc<-,polyPdmpModel-method}
\alias{ratefunc<-,polyPdmpModel-method}
\alias{ratesprays<-,polyPdmpModel-method}
\alias{dynsprays<-,polyPdmpModel-method}
\title{Accessor functions for Class polyPdmpModel}
\usage{
dynpolys(obj)

dynpolys(obj) <- value

ratepolys(obj)

ratepolys(obj) <- value

ratesprays(obj)

ratesprays(obj) <- value

dynsprays(obj)

dynsprays(obj) <- value

\S4method{dynpolys}{polyPdmpModel}(obj)

\S4method{ratepolys}{polyPdmpModel}(obj)

\S4method{ratesprays}{polyPdmpModel}(obj)

\S4method{dynsprays}{polyPdmpModel}(obj)

\S4method{dynpolys}{polyPdmpModel}(obj) <- value

\S4method{ratepolys}{polyPdmpModel}(obj) <- value

\S4method{parms}{polyPdmpModel}(obj) <- value

\S4method{dynfunc}{polyPdmpModel}(obj) <- value

\S4method{ratefunc}{polyPdmpModel}(obj) <- value

\S4method{ratesprays}{polyPdmpModel}(obj) <- value

\S4method{dynsprays}{polyPdmpModel}(obj) <- value
}
\arguments{
\item{obj}{an object of class \code{\link{polyPdmpModel}}}

\item{value}{the value that shall be set}
}
\description{
Get or set the slot values for an object of class
\code{\link{polyPdmpModel}}. There are no methods for setting new values of
the slots \code{dynsprays}, \code{dynfunc}, \code{ratesprays} and 
\code{ratefunc}. Their value can be changed by setting a new value to
\code{dynpolys}, \code{ratepolys} or \code{parms}. Only these slots
should be accessed by the user. Slot \code{out} will be set to NULL, if
new values for the other slots are provided.
}
\details{
It is also possible to modify the slots directly,
e.g. slot \code{dynsprays} of a model \code{model} with \code{model@dynsprays}, 
but this is not recommended! The slots have to fit together and there is
no method warning you of possible misleading results.
}
\note{
only works for one discrete variable

only works for one discrete variable
}
