% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supports.R
\name{support}
\alias{support}
\alias{getSupport}
\alias{supports}
\alias{plotSupport}
\alias{getsupport}
\alias{plotsupport}
\title{Return the support for gene regulation models}
\usage{
getSupport(model)

plotSupport(ggplot = NULL, support)
}
\arguments{
\item{model}{Object of class \code{pdmpModel} or \code{polyPdmp}. This has to
be one of the previous mentioned models.}

\item{ggplot}{Object of class ggplot. The bounds will be plotted onto this 
object. If NULL, a new plot object will be created.}

\item{support}{data.frame that has the same structure as the data.frame
returned by \code{getSupport}.}
}
\description{
In PROM, the support of the following gene regualtion models has been
calculated: \code{[Model K][modelK]}, \code{[Model K2][modelK2]},
\code{[Model F][modelF]}, \code{[Model KF][modelKF]}, 
\code{[Model BF][modelBF]} and \code{[Model T][modelT]}.
Method \code{getSupport} returns a data.frame with the lower and upper bounds
of the support for every time value, method \code{plotSupport} adds these
bounds as lines into a plot given as \code{ggplot} object.
}
\note{
Method \code{plotSupport} currently only works for models with one
continous variable.
}
\examples{
data(genePolyT)
getSupport(genePolyT) \%>\% head

data(genePdmpBF)
ms <-  multSim(genePdmpBF, seeds = 1:4)
plotSupport(ggplot = plotSeeds(ms, 1:4), data = getSupport(genePdmpBF))
}
